third_party/blink/web_tests/media/media-fragments/TC0001-expected.txt


Title: TC0001
Fragment: 't=,'
Comment: Syntax error, not allowed according to the ABNF. The media fragment is ignored.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
END OF TEST
